requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.3
Brotli==1.0.9
build==0.10.0
CacheControl==0.12.11
camelot-py==0.9.0
certifi==2023.5.7
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
cryptography==42.0.5
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
dj-database-url==2.0.0
Django==4.1.7
django-ckeditor==6.5.1
django-classy-tags==4.0.0
django-cms==3.11.1
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-environ==0.10.0
django-filter==23.1
django-formtools==2.4
django-import-export==3.3.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-sekizai==4.0.0
django-treebeard==4.6.1
djangocms-admin-style==3.2.3
djangorestframework==3.14.0
docopt==0.6.2
dulwich==0.21.5
et-xmlfile==1.1.0
filelock==3.12.0
frozenlist==1.3.3
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
installer==0.7.0
jaraco.classes==3.2.3
JPype1==1.5.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
MarkupPy==1.14
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
newsapi==0.1.1
newsapi-python==0.2.7
numpy==1.25.2
odfpy==1.4.1
openai==0.27.6
opencv-python==4.8.0.76
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pdfminer.six==20231228
pexpect==4.8.0
Pillow==9.5.0
pipreqs==0.4.13
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycountry==22.3.5
pycparser==2.21
pygame==2.5.2
PyPDF2==3.0.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.0
PyYAML==6.0.1
rapidfuzz==2.15.1
requests==2.30.0
requests-toolbelt==0.10.1
shell==1.0.1
shellingham==1.5.0.post1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.3
tablib==3.5.0
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
trove-classifiers==2023.5.2
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.18
virtualenv==20.21.1
webencodings==0.5.1
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
